backend/requirements.txt
Flask
Flask-SocketIO==5.3.1
python-engineio==4.3.4
python-socketio==5.0.2
SQLAlchemy
requests
Flask-CORS
bidict
python-dotenv
PyJWT
gevent
greenlet
simple-websocket
googletrans==4.0.0rc1
gevent-websocket
